+++ examples/collection/assembly/index.ts
@@ -1,6 +1,6 @@
 import { collections } from "@hypermode/functions-as";
 import { models } from "@hypermode/functions-as";
-import { EmbeddingsModel } from "@hypermode/models-as/models/openai/embeddings";
+import { OpenAIEmbeddingsModel } from "@hypermode/models-as/models/openai/embeddings";
 
 // These names should match the ones defined in the hypermode.json manifest file.
 const modelName: string = "embeddings";
@@ -8,19 +8,27 @@ const myProducts: string = "myProducts";
 const searchMethods: string[] = ["searchMethod1", "searchMethod2"];
 
 // This function takes input text and returns the vector embedding for that text.
-export function embed(text: string): f64[] {
-  const model = models.getModel<EmbeddingsModel>(modelName);
+export function embed(text: string[]): f32[][] {
+  const model = models.getModel<OpenAIEmbeddingsModel>(modelName);
   const input = model.createInput(text);
   const output = model.invoke(input);
-  return output.data[0].embedding;
+  return output.data.map<f32[]>((d) => d.embedding);
 }
 
-export function addProduct(description: string): string {
+export function addProduct(description: string): string[] {
   const response = collections.upsert(myProducts, null, description);
   if (!response.isSuccessful) {
     throw new Error(response.error);
   }
-  return response.key;
+  return response.keys;
+}
+
+export function addProducts(descriptions: string[]): string[] {
+  const response = collections.upsertBatch(myProducts, [], descriptions);
+  if (!response.isSuccessful) {
+    throw new Error(response.error);
+  }
+  return response.keys;
 }
 
 export function deleteProduct(key: string): string {